refactor(data): place optimistic helpers in lib (#48)
